requirements.txt
certifi==2023.11.17
cfgv==3.3.1
charset-normalizer==3.3.2
distlib==0.3.8
dnspython==1.16.0
filelock==3.12.2
identify==2.5.24
idna==3.6
importlib-metadata==6.7.0
mongoengine==0.27.0
nodeenv==1.8.0
platformdirs==4.0.0
pre-commit==2.21.0
pymongo==4.6.1
python-dotenv==0.21.1
PyYAML==6.0.1
requests==2.31.0
typing-extensions==4.7.1
urllib3==2.0.7
virtualenv==20.25.0
zipp==3.15.0
